% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadmatrix.R
\name{quadmatrix}
\alias{quadmatrix}
\title{Solving Quadratic Matrix Equations}
\usage{
quadmatrix(A, B, C)
}
\arguments{
\item{A}{A Matrix}

\item{B}{A Matrix}

\item{C}{A Matrix}
}
\value{
Matrix X
}
\description{
Given inputs A,B and C, this package solves the matrix equation A*X^2 - B*X - C = 0.
}
\examples{
B = matrix(c(1,3,3,1),nrow = 2,ncol = 2)
A = matrix(c(1.5,2,2,1.5),nrow = 2,ncol = 2)
C = matrix(c(1.5,1,1,1.5),nrow = 2,ncol = 2)
quadmatrix(A,B,C)



}
